refactor: EnableExtension 내부 클래스 외부로 이동
